ADDED A NEW POST CREATOR (PYTHON)
This is actualy my first commit which has a comment. So, what I did was add a python scrip that allows me to input data into a GUI which later gets turned into a html file. It is also going to be designed to make sure that the website javascript variables are updated to link with the .html file. I believe that I should finish it by tommorow. -Artur